Rename svg files and imports
